width

+++ vortex-duckdb/src/duckdb/connection.rs
@@ -188,7 +188,7 @@ mod tests {
 
         let chunk = result.into_iter().next().unwrap();
         let vec = chunk.get_vector(0);
-        let slice = vec.as_slice_with_len::<i64>(chunk.len().as_());
+        let slice = vec.as_slice_with_len::<i32>(chunk.len().as_());
 
         let mut vec_str = chunk.get_vector(1);
         let slice_str = unsafe { vec_str.as_slice_mut::<duckdb_string_t>(chunk.len().as_()) };
